Update the key words generation
The main change is that now we process the document context once and the then we process each chunk and in the end all of them are added to the TF-IDF model.
Previously, we initialize the model over and over again and that adds computational time.

The fallback methods are the same as the previous approach.